+++ addOns/network/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update dependencies.
 
+### Fixed
+- Correct the generation of server certificates for domain names with more than 64 characters.
+
 ## [0.28.0] - 2026-05-21
 ### Changed
 - Update dependencies (Issue 9337).